% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_color_map.R
\name{expand_regions}
\alias{expand_regions}
\title{Expand regions to a desired width}
\usage{
expand_regions(zu_obj, sizes, both = 0, left = 0, right = 0,
  strand = TRUE, trim = TRUE)
}
\arguments{
\item{zu_obj}{A ZentUtils object}

\item{sizes}{Path to a two-column file of chromosome sizes from UCSC (can be a URL)}

\item{both}{Distance to add to the start and end of each region}

\item{left}{Distance to add to the start of the region}

\item{right}{Distance to add to the end of the region}

\item{strand}{Define left and right based on strand}

\item{trim}{Adjust coordinates for out-of-bounds intervals}
}
\value{
A GRanges object in slot 'expanded_regions'
}
\description{
Expand regions to a desired width
}
\examples{
expand_regions(reb1,
sizes = "https://hgdownload.cse.ucsc.edu/goldenPath/sacCer3/bigZips/sacCer3.chrom.sizes", both = 50)
}
