% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_color_map.R
\name{expand_summits}
\alias{expand_summits}
\title{Expand peak summits to a desired width}
\usage{
expand_summits(zu_obj, length = 100)
}
\arguments{
\item{zu_obj}{A ZentUtils object}

\item{length}{Number of bases to expand the summit}
}
\value{
A GRanges object in slot 'regions' (overwrites original regions)
}
\description{
Expand peak summits to a desired width
}
\examples{
expand_summits(zu_obj, length = 100)
}
