% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_color_map.R
\name{get_seqs}
\alias{get_seqs}
\title{Retrieve sequences corresponding to ranges of interest}
\usage{
get_seqs(zu_obj, region_type = "imported", genome)
}
\arguments{
\item{zu_obj}{A ZentTools object}

\item{region_type}{Whether to extract sequences from original regions
("imported") or expanded regions ("exported")}

\item{genome}{A BSgenome object for the organism of interest}
}
\value{
A DNAStringSet object in slot 'seqs'
}
\description{
Retrieve sequences corresponding to ranges of interest
}
\examples{
saccer3 <- BSgenome.Scerevisiae.UCSC.sacCer3
get_seq(zu_obj, genome = saccer3)
}
