% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_functions.R
\name{sort_regions}
\alias{sort_regions}
\title{Sort Regions}
\usage{
sort_regions(zu_obj, by = "coord", decreasing = FALSE)
}
\arguments{
\item{zu_obj}{A ZentUtils object.}

\item{by}{Sort by chromosome + coordinates + start ("coord") or column 5
of the imported ranges (automatically named "score" upon import).}

\item{decreasing}{Whether the GRanges object should be sorted in descending order.}
}
\value{
A sorted GRanges object in the 'regions' slot of the ZentUtils object
(overwrites original region order).
}
\description{
Sorts imported regions by chromosomal location or score.
}
\examples{
zent <- sort_regions(zent, by = "coord", decreasing = FALSE)
}
