% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zu_obj.R
\docType{class}
\name{zu_obj-class}
\alias{zu_obj-class}
\title{ZentUtils class definition}
\description{
ZentUtils class definition
}
\section{Slots}{

\describe{
\item{\code{regions}}{A GRanges object containing imported regions.}

\item{\code{expanded_regions}}{A GRanges object containing expanded regions.}

\item{\code{seqs}}{A DNAStringSet object containing sequences of regions of interest.}
}}

