% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zu_obj.R
\name{zentutils}
\alias{zentutils}
\title{ZentUtils constructor function}
\usage{
zentutils(data)
}
\arguments{
\item{data}{A BED file containing the regions of interest}
}
\value{
A ZentUtils object
}
\description{
ZentUtils constructor function
}
\examples{
zentutils("~/CTCF_peaks.bed")
}
