% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_color_map.R
\name{color_map}
\alias{color_map}
\title{Generate sequence color map}
\usage{
color_map(zu_obj, cols = NA)
}
\arguments{
\item{zu_obj}{A ZentUtils object}

\item{cols}{Colors to use for base representation. Use NA for default colors
or provide a character vector of length 4.}
}
\value{
A sequence color map
}
\description{
Generate sequence color map
}
\examples{
color_map(zent, cols = c("green", "blue", "yellow", "red"))
}
