% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_color_map.R
\name{expand_regions}
\alias{expand_regions}
\title{Expand regions to a desired width}
\usage{
expand_regions(zu_obj, genome, length = 100)
}
\arguments{
\item{zu_obj}{A ZentUtils object}

\item{genome}{A BSgenome object for the organism of interest}

\item{length}{Number of bases to expand the region}
}
\value{
A GRanges object in slot 'expanded_regions'
}
\description{
Expand regions to a desired width
}
\examples{
genome <- BSgenome.Scerevisiae.UCSC.sacCer3
zent <- expand_regions(zent, genome, 100)
}
