% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_functions.R
\name{expand_regions}
\alias{expand_regions}
\title{Expand Regions}
\usage{
expand_regions(zu_obj, length = 10)
}
\arguments{
\item{zu_obj}{A ZentUtils object.}

\item{length}{Number of bp to add to both sides of each region.}
}
\value{
A GRanges object in the 'expanded_regions' slot of the ZentUtils object.
}
\description{
Expands regions to a desired width. Sequences that are partially out-of-bounds
after expansion are removed.
}
\examples{
zent <- expand_regions(zent, length = 10)
}
