% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zu_obj.R
\name{zentutils}
\alias{zentutils}
\title{ZentUtils Constructor Function}
\usage{
zentutils(data, genome = NA, mito_chr = "chrM")
}
\arguments{
\item{data}{A BED file containing the regions of interest.}

\item{genome}{Identifier for the genome from which the regions were derived.}

\item{mito_chr}{Seqname of mitochondrial chromosome.}
}
\value{
A new ZentUtils object with a GRanges object in the 'regions' slot.
}
\description{
Creates a new ZentUtils object that stores a GRanges object generated from
an imported BED file.
}
\details{
\itemize{
\item{The BED file should contain 6 columns, with chromosome, start, and
end coordinates in columns 1-3 and strand in column 6.}
\item{BED columns names are converted to "chrom", "start", "end",
"region_name", "score", and "strand" upon import.}
}
}
\examples{
zent <- zentutils(system.file("extdata", "homer_reb1_badis_motif_scan.bed", package = "ZentUtils"), genome = "sacCer3", mito_chr = "chrM")
}
